オドメトリ関連を整理/再設計. (motion) MotionのオドメトリAPIを再設計. 運動学によるオドメトリ計算を追加. HeadYawとHeadPitchの角速度による制御APIを追加. (localization) OdometryLocalizationを追加. updateOdometryによる更新を修正. (sensation) SensoryCortexを整理 (misc) BooleanFilterを追加. rpy表現の順序を変更.

+++ src/jp/ac/fit/asura/nao/event/MotionEventListener.java
@@ -9,14 +9,14 @@ import jp.ac.fit.asura.nao.motion.Motion;
 
 /**
  * @author sey
- * 
+ *
  * @version $Id: MotionEventListener.java 627 2008-06-26 18:13:39Z sey $
- * 
+ *
  */
 public interface MotionEventListener extends EventListener {
 	public void updatePosture();
 
-	public void updateOdometry(int forward, int left, float turnCCW);
+	public void updateOdometry(float forward, float left, float turnCCW);
 
 	public void startMotion(Motion motion);
